smart_chat_system/requirements.txt

# স্মার্ট চ্যাট সিস্টেমের জন্য প্রয়োজনীয় প্যাকেজ

# ওয়েব ফ্রেমওয়ার্ক
Flask==2.3.0
Flask-SocketIO==5.3.0

# ইমেজ প্রসেসিং
Pillow==9.5.0

# সাপোর্টিং লাইব্রেরি
python-socketio==5.8.0
python-engineio==4.7.1
eventlet==0.33.3
